cleanup(app): drop @takazudo link: deps, env-var prefixes, public-copy plugin half
- app/package.json: remove @takazudo/zfb and @takazudo/zfb-runtime link:
  dependencies (now embedded in zfb binary via upstream fixes #183/#198);
  simplify build script from darwin-arm64-specific ZFB_ESBUILD_BIN/
  ZFB_TAILWIND_BIN prefixes to plain `zfb build`
- app/.env.build: delete (contained darwin-arm64-specific env var stubs)
- app/plugins/copy-public.mjs: delete (public-copy half made obsolete by
  zfb upstream fix #192 which now copies public/ natively)
- app/plugins/rename-shell.mjs: add (shell-rename half only — zfb router
  convention workaround still needed until upstream adds underscore-page
  escape hatch)
- app/zfb.config.ts: update plugin reference from copy-public.mjs to
  rename-shell.mjs; update comment to reflect single remaining purpose

Note: full no-node build (without any app/node_modules) waits for Wave 3
(#31 drops preact dep); the embedded @takazudo extraction only kicks in when
no node_modules exists at all — a zfb-side gap to address before Wave 3.
