--- app/.env.build
@@ -1,9 +1,0 @@
-# Build-time environment variables for zfb.
-# Source this before running pnpm zfb build:
-#   source .env.build && pnpm zfb build
-#
-# ZFB_ESBUILD_BIN: path to esbuild 0.25.12 standalone CLI
-# ZFB_TAILWIND_BIN: path to tailwindcss v4 standalone CLI
-
-ZFB_ESBUILD_BIN="$HOME/repos/myoss/zfb/node_modules/.pnpm/@esbuild+darwin-arm64@0.25.12/node_modules/@esbuild/darwin-arm64/bin/esbuild"
-ZFB_TAILWIND_BIN="$HOME/repos/myoss/zfb/crates/zfb/binaries/tailwindcss-v4"
