+++ app/zfb.config.ts
@@ -8,9 +8,9 @@ export default defineConfig({
   // No SSR adapter — pure static build; axum serves the static output at runtime.
   plugins: [
     {
-      // Workaround: zfb does not copy public/ to dist/ in production builds.
-      // This plugin fills the gap. Remove when zfb adds native public-dir copy.
-      name: "./plugins/copy-public.mjs",
+      // Workaround: zfb router skips underscore-prefixed pages; rename dist/shell/ → dist/_shell/.
+      // Remove when zfb supports an opt-in escape hatch for underscore-prefixed pages.
+      name: "./plugins/rename-shell.mjs",
     },
   ],
 });
